split models into submodule
