+++ app/vodloader/models/__init__.py
@@ -1,0 +1,28 @@
+from base import *
+from TwitchChannel import *
+from TwitchChannelUpdate import *
+from TwitchStream import *
+from VideoFile import *
+from YouTubeVideo import *
+from TwitchClient import *
+from TwitchAuth import *
+from ClearChatEvent import *
+from ClearMsgEvent import *
+
+
+MODELS: List[BaseModel] = [
+    TwitchChannel,
+    TwitchStream,
+    TwitchChannelUpdate,
+    VideoFile,
+    YoutubeVideo,
+    TwitchClient,
+    TwitchAuth,
+    Message,
+    ClearChatEvent,
+    ClearMsgEvent,
+]
+
+async def initialize_models():
+    for model in MODELS:
+        await model.initialize()
